#1819 Move OMPTargetTrans to psyclone.psyir.transformations and improve docstrings
