+++ src/psyclone/psyir/transformations/__init__.py
@@ -67,6 +67,7 @@ from psyclone.psyir.transformations.loop_tiling_2d_trans \
     import LoopTiling2DTrans
 from psyclone.psyir.transformations.loop_trans import LoopTrans
 from psyclone.psyir.transformations.nan_test_trans import NanTestTrans
+from psyclone.psyir.transformations.omp_target_trans import OMPTargetTrans
 from psyclone.psyir.transformations.omp_taskwait_trans import OMPTaskwaitTrans
 from psyclone.psyir.transformations.profile_trans import ProfileTrans
 from psyclone.psyir.transformations.psy_data_trans import PSyDataTrans
@@ -100,6 +101,7 @@ __all__ = ['ArrayRange2LoopTrans',
            'LoopTrans',
            'NanTestTrans',
            'OMPTaskwaitTrans',
+           'OMPTargetTrans',
            'ProfileTrans',
            'PSyDataTrans',
            'ReadOnlyVerifyTrans',
